moved project related pages to /pages
