--- src/components/generics/Project/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Project'
