+++ src/entities/Project.ts
@@ -1,0 +1,6 @@
+export interface Project {
+    name: string,
+    description: string,
+    members: string[],
+    isActive: boolean,
+}
