+++ src/pages/Resources/index.ts
