+++ src/services/adapters/projectAdapter.ts
@@ -1,6 +1,6 @@
 import type { BaseResponse } from './types'
 import projectAPI from '../api/projectAPI'
-import { Project } from '@components/generics/Project/Project'
+import { Project } from '@entities/Project'
 
 interface ProjectResponse extends BaseResponse {
     data? : Project[]
@@ -9,6 +9,7 @@ interface ProjectResponse extends BaseResponse {
 export const callGetAllProjectsAPI = async (): Promise<ProjectResponse> => {
     try {
         const res = await projectAPI.getAllProjects()
+        console.log(res.data);
         return res.data as ProjectResponse;
     } catch (err: any) {
         return err.response.data as BaseResponse;
